Increase default max current for steppers
Turns out all of the Steppers of the Ender 3X are rated for max 1500mA.
I'm not sure why a value of 580 was set but this causes the Z axis to
skip when going up, due to the additional force required to move the arm
up.

I see no reason to set the value this low however I'm not 100% sure I've
got the values right. This is a compromise until I get some more
experiance printing with these values.